lexer -> location
