+++ react/src/urban-stats-script/interpreter.ts
@@ -3,7 +3,7 @@ import { assert } from '../utils/defensive'
 import { UrbanStatsASTStatement, UrbanStatsASTExpression, UrbanStatsASTLHS, UrbanStatsASTArg, locationOf, unify } from './ast'
 import { Context } from './context'
 import { addAdditionalDims, broadcastApply, broadcastCall } from './forward-broadcasting'
-import { LocInfo } from './lexer'
+import { LocInfo } from './location'
 import { expressionOperatorMap } from './operators'
 import { splitMask } from './split-broadcasting'
 import { renderType, unifyType, USSRawValue, USSType, USSValue, USSVectorType, ValueArg, undocValue, canUnifyTo } from './types-values'
