+++ react/src/urban-stats-script/lexer.ts
@@ -1,5 +1,6 @@
 import { assert } from '../utils/defensive'
 
+import { Block, LocInfo } from './location'
 import { expressionOperatorMap } from './operators'
 
 const operators = [...expressionOperatorMap.keys(), '=', ',', ';', '.', ':']
@@ -17,23 +18,6 @@ interface BracketToken { type: 'bracket', value: '(' | ')' | '{' | '}' | '[' | '
 interface ErrorToken { type: 'error', value: string }
 type NonErrorToken = NumericToken | IdentifierToken | KeywordToken | StringToken | OperatorToken | BracketToken
 type Token = NonErrorToken | ErrorToken
-
-export type Block = { type: 'single', ident: string } | { type: 'multi' }
-
-export interface SingleLocationWithinBlock {
-    lineIdx: number
-    colIdx: number
-    charIdx: number
-}
-
-export type SingleLocation = SingleLocationWithinBlock & { block: Block }
-
-export interface LocInfo {
-    start: SingleLocation
-    end: SingleLocation
-}
-
-export const noLocation = { start: { block: { type: 'multi' }, lineIdx: 0, colIdx: 0, charIdx: 0 }, end: { block: { type: 'multi' }, lineIdx: 0, colIdx: 0, charIdx: 0 } } satisfies LocInfo
 
 export interface AnnotatedToken {
     token: Token
