+++ react/unit/urban-stats-script-parsing.test.ts
@@ -2,7 +2,9 @@ import assert from 'assert/strict'
 import fs from 'node:fs'
 import { test } from 'node:test'
 
-import { Block, lex, noLocation } from '../src/urban-stats-script/lexer'
+import { lex } from '../src/urban-stats-script/lexer'
+import { noLocation } from '../src/urban-stats-script/location'
+import { Block } from '../src/urban-stats-script/location'
 import { allIdentifiers, parse, parseNoErrorAsCustomNode, toSExp, unparse } from '../src/urban-stats-script/parser'
 
 import { emptyContext, parseProgram } from './urban-stats-script-utils'
